Support SPI, BMA253, more features
This is a rewrite of a significant part of the library:

I moved the I2C communication code to interface.rs, abstracted over it
and added support for SPI communication. I also added support for the
bma253, which I'm currently working with. I created another trait to
abstract over the details of each bma2xx. I also created register.rs
which lists all registers described in the bma253 data sheet, using the
names from that data sheet so they are easy to look up. Finally, I
implemented configuration of the tap sensing feature.

I haven't tested these changes with the bma222e or with I2C as I don't
have the hardware for that, but I have mostly left that code alone so I
still expect it to work. There are some breaking changes in how to
create an instance of the device, see the changes to the documentation
for that.